requirements.txt
PyQt5
scapy
joblib
pandas
pygame
psutil
python-nmap
requests
beautifulsoup4
reportlab
opencv-python